Restore old MV2 path
