+++ src/manifest.json
@@ -62,8 +62,5 @@
         "switchEngine": {
             "description": "__MSG_theme_generation_mode__"
         }
-    },
-    "web_accessible_resources": [
-        "inject/proxy.js"
-    ]
+    }
 }
