+++ tasks/bundle-js.js
@@ -46,8 +46,9 @@ function patchMV3JS(/** @type {string} */code) {
  * @property {string} src
  * @property {string} dest
  * @property {string} reloadType
- * @property {({debug}) => Promise<void>} postBuild
+ * @property {(({debug}) => Promise<void>) | undefined} postBuild
  * @property {string[]} watchFiles
+ * @property {(typeof PLATFORM.CHROME) | undefined} platform
  */
 
 /** @type {JSEntry[]} */
@@ -81,12 +82,10 @@ const jsEntries = [
         watchFiles: null,
     },
     {
-        src: 'src/inject/dynamic-theme/stylesheet-proxy.ts',
+        src: 'src/inject/dynamic-theme/stylesheet-proxy-mv3-injector.ts',
         dest: 'inject/proxy.js',
         reloadType: reload.FULL,
-        async postBuild({debug}) {
-            await copyToBrowsers({cwdPath: this.dest, debug});
-        },
+        platform: PLATFORM.CHROME_MV3,
         watchFiles: null,
     },
     {
@@ -155,12 +154,12 @@ async function bundleJS(/** @type {JSEntry} */entry, {debug, watch}) {
     });
     entry.watchFiles = bundle.watchFiles;
     await bundle.write({
-        file: `${getDestDir({debug, platform: PLATFORM.CHROME})}/${dest}`,
+        file: `${getDestDir({debug, platform: entry.platform || PLATFORM.CHROME})}/${dest}`,
         strict: true,
         format: 'iife',
         sourcemap: debug ? 'inline' : false,
     });
-    await entry.postBuild({debug});
+    entry.postBuild && await entry.postBuild({debug});
 }
 
 function getWatchFiles() {
